I realy enjoyed using terminal_typing_tutor, so I made some (IMHO) improvements:
* When the user hits the correct key, change the color
to blue (from the original green). I am colorblind
and the green was very difficult to differentiate.

* Changed the behavior of multiline lessons so that
the user can press space to get to the next line instead
of return. I thought this was more intuitive behavior.
Blank lines are also entered as space, but I don't think
this takes away from learning to type because there
isn't much value in learning to type the return key!

* Added Quote of the Day (QOTD) series. I thought it
would be helpful to practice typing somewhat random
text. All the material comes from wikiquote.org and
requires installation via 'pip install wikiquote'.
The first lesson simply gets the quote of the day from
the front page of wikiquote.  The second lesson picks
a random 'topic' or 'title' and then gets up to 20 quotes.
The random topic and quotes are updated each time the
user selects this lesson, so a variety of text can be
practiced. The third lesson gets up to 30 quotes where
the topic, or title, is Thomas Edison (my inspiration
for using quotes as typing lessons). Ultimately, I would
also like to have a fourth lesson where the user can
submit a topic to fetch quotes for.

Using the wikiquote API introduced characters that were
not easily typeable (e.g., elipsis, hyphen). They were
'replaced' with the closest ASCII characters. For other
untypeable characters, I replaced them with '***' to
indicate a replacement, and also to be able to proceed
with the lesson.